Ordered fills fixed

+++ ordenacao.h
@@ -4,13 +4,13 @@ typedef struct __result {
     char nome[30];
 } RESULTADOS;
 
-void copiar(unsigned long * , unsigned long * , unsigned long);
-void mostrar(unsigned long * , unsigned long);
+void copiar(unsigned long long * , unsigned long long * , unsigned long long);
+void mostrar(unsigned long long * , unsigned long long);
 void mostrar_resultados(RESULTADOS);
 
-void bubble(unsigned long * , unsigned long, RESULTADOS * );
-void bubble_otimizado(unsigned long * , unsigned long, RESULTADOS * );
-void select_sort(unsigned long *, unsigned long, RESULTADOS *);
-void insert_sort(unsigned long *, unsigned long, RESULTADOS *);
-void quick_sort(unsigned long *, unsigned long, RESULTADOS *);
-void __quick_sort(unsigned long *, unsigned long, unsigned long, RESULTADOS *);
+void bubble(unsigned long long * , unsigned long long, RESULTADOS * );
+void bubble_otimizado(unsigned long long * , unsigned long long, RESULTADOS * );
+void select_sort(unsigned long long *, unsigned long long, RESULTADOS *);
+void insert_sort(unsigned long long *, unsigned long long, RESULTADOS *);
+void quick_sort(unsigned long long *, unsigned long long, RESULTADOS *);
+void __quick_sort(unsigned long long *, unsigned long long, unsigned long long, RESULTADOS *);
